Organized the functions with express
