+++ functions/package.json
@@ -14,12 +14,16 @@
   },
   "main": "lib/index.js",
   "dependencies": {
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
     "faker": "^5.1.0",
     "firebase-admin": "^8.10.0",
     "firebase-functions": "^3.6.1",
     "fp-ts": "^2.8.4"
   },
   "devDependencies": {
+    "@types/cors": "^2.8.8",
+    "@types/express": "^4.17.8",
     "@types/faker": "^5.1.2",
     "firebase-functions-test": "^0.2.0",
     "tslint": "^5.12.0",
