+++ functions/src/api/cards/index.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import {postBuyCards} from "./api/postBuyCards";
+
+const cards = Router();
+
+cards.post('/buy', postBuyCards);
+
+export default cards;
