+++ functions/src/api/lobby/index.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import {postCreateLobby} from "./api/postCreateLobby";
+
+const lobby = Router();
+
+lobby.post('/', postCreateLobby);
+
+export default lobby;
